requirements.txt
discord.py==1.7.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
chardet==4.0.0
multidict==5.1.0
attrs==21.2.0
idna==3.2
typing-extensions==3.10.0.0
yarl==1.6.3
colorama==0.4.6
